move "labeling.go" helpers from core to cmd/label.
